Move `ApiDto`, update dependencies
