+++ api/src/main/java/com/codeforcommunity/dto/ApiDto.java
@@ -1,16 +1,9 @@
-package com.codeforcommunity.api;
+package com.codeforcommunity.dto;
 
 import com.codeforcommunity.exceptions.HandledException;
 import com.codeforcommunity.exceptions.MalformedParameterException;
 import java.util.List;
 
-/**
- * An abstract class that all incoming request DTOs must extend. This is used to validate the
- * objects before they're used to ensure that certain fields are not-null/not-empty or fulfill other
- * criteria.
- *
- * @author Conner Nilsen
- */
 public abstract class ApiDto {
   /**
    * Verify if the extending DTO is a valid object.
