+++ api/src/main/java/com/codeforcommunity/dto/auth/NewUserRequest.java
@@ -1,6 +1,6 @@
 package com.codeforcommunity.dto.auth;
 
-import com.codeforcommunity.api.ApiDto;
+import com.codeforcommunity.dto.ApiDto;
 import com.codeforcommunity.dto.IDTO;
 import java.util.ArrayList;
 import java.util.List;
