+++ api/src/main/java/com/codeforcommunity/rest/RestFunctions.java
@@ -1,6 +1,6 @@
 package com.codeforcommunity.rest;
 
-import com.codeforcommunity.api.ApiDto;
+import com.codeforcommunity.dto.ApiDto;
 import com.codeforcommunity.exceptions.MalformedParameterException;
 import com.codeforcommunity.exceptions.MissingHeaderException;
 import com.codeforcommunity.exceptions.MissingParameterException;
