+++ pom.xml
@@ -18,7 +18,6 @@
   <properties>
     <maven.compiler.source>1.8</maven.compiler.source>
     <maven.compiler.target>1.8</maven.compiler.target>
-    <vertx.version>3.9.1</vertx.version>
   </properties>
 
   <build>
@@ -146,7 +145,7 @@
       <dependency>
         <groupId>com.fasterxml.jackson.core</groupId>
         <artifactId>jackson-core</artifactId>
-        <version>2.9.10</version>
+        <version>2.11.3</version>
       </dependency>
 
       <!-- Used for JWT handling -->
