+++ service/pom.xml
@@ -87,7 +87,6 @@
     <dependency>
       <groupId>com.fasterxml.jackson.core</groupId>
       <artifactId>jackson-core</artifactId>
-      <version>2.9.10</version>
     </dependency>
     <dependency>
       <groupId>org.junit.jupiter</groupId>
